requirements.txt
streamlit==1.32.0
streamlit-folium==0.15.1
folium==0.15.1
pandas==2.2.0
requests==2.31.0
python-dotenv==1.0.1
pytest==7.4.0
pytest-cov==4.1.0
boto3==1.34.0
schedule==1.2.1